requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv
pymysql
flask-jwt-extended
Flask-Login
Flask-wtf
flask-bcrypt
flask_migrate
